Implement and setup polyversion script.
The script represents an alternative `sphinx-multiversion` and builds multiple versions
of the documentation which are sourced from the git repo.
The script was created since it supports more features than
sphinx-multiversion including

- Concurrent builds
- Easy integration with Github Actions including caching
- Extended configuration options
- Poetry integrations + isolated builds
- Generate global, version independent, toplevel files

Additionally  `sphinx-multiversion` is no longer maintained.

This also updates existing files for use with the new script and
adds some files that will be placed in the root of the polyversion build.

The script depends on `jinja2` which is already an indirekt dependency
through `sphinx`.

* bin/sphinx-polyversion/sphinx_polyversion.py : Implement.

* bin/sphinx-polyversion/pyproject.toml: Package config and jinja2 dependency.

* pyproject.toml : Add `polyversion` to group `docs`.

* poetry.lock : Autogenerate through `pre-commit`.

* docs/source/conf.py : Adjust to new script. Fix some more stuff.

* docs/source/_templates/versioning.html : Adjust to new script.

* docs/source/_polyversion/templates/index.html : Added.
    Its a template for the root file of the documentation redirecting
    to the latest tags root.

* docs/source/_polyversion/static/.nojekyll : Added.
    Its a marker file for github-pages.

* Makefile : Add make phony for polyversion script.